Combine CombatNotifier and IdleNotifier plugins
