--- runelite-client/src/main/java/net/runelite/client/plugins/combatnotifier/CombatNotifier.java
@@ -1,95 +1,0 @@
-/*
- * Copyright (c) 2017, Kronos <https://github.com/KronosDesign>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package net.runelite.client.plugins.combatnotifier;
-
-import com.google.common.eventbus.Subscribe;
-import java.time.Duration;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
-import net.runelite.api.Actor;
-import net.runelite.api.Client;
-import net.runelite.api.GameState;
-import net.runelite.api.Player;
-import net.runelite.client.RuneLite;
-import net.runelite.client.events.GameStateChanged;
-import net.runelite.client.plugins.Plugin;
-import net.runelite.client.plugins.PluginDescriptor;
-import net.runelite.client.task.Schedule;
-
-@PluginDescriptor(
-	name = "Combat notifier"
-)
-public class CombatNotifier extends Plugin
-{
-	private final Client client = RuneLite.getClient();
-	private final RuneLite runelite = RuneLite.getRunelite();
-	private final CombatNotifierConfig config = runelite.getConfigManager()
-		.getConfig(CombatNotifierConfig.class);
-
-	private Instant lastInteracting;
-
-	@Override
-	protected void startUp() throws Exception
-	{
-	}
-
-	@Override
-	protected void shutDown() throws Exception
-	{
-	}
-
-	@Subscribe
-	public void onGameStateChanged(GameStateChanged gameStateChanged)
-	{
-		lastInteracting = null;
-	}
-
-	@Schedule(
-		period = 1,
-		unit = ChronoUnit.SECONDS
-	)
-	public void checkIdle()
-	{
-		if (client.getGameState() != GameState.LOGGED_IN || !config.isEnabled())
-		{
-			return;
-		}
-
-		Player local = client.getLocalPlayer();
-		Actor opponent = local.getInteracting();
-		if (opponent != null && opponent.getCombatLevel() > 0)
-		{
-			lastInteracting = Instant.now();
-		}
-
-		Duration waitDuration = Duration.ofMillis(config.getTimeout());
-		if (lastInteracting != null && Instant.now().compareTo(lastInteracting.plus(waitDuration)) >= 0)
-		{
-			runelite.notify("[" + local.getName() + "] is now out of combat!");
-			lastInteracting = null;
-		}
-	}
-
-}
