+++ runelite-client/src/main/java/net/runelite/client/plugins/idlenotifier/IdleNotifierConfig.java
@@ -1,15 +1,15 @@
 /*
- * Copyright (c) 2017, Kronos <https://github.com/KronosDesign>
+ * Copyright (c) 2017, Devin French <https://github.com/devinfrench>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *
  * 1. Redistributions of source code must retain the above copyright notice, this
- *     list of conditions and the following disclaimer.
+ *    list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
- *     this list of conditions and the following disclaimer in the documentation
- *     and/or other materials provided with the distribution.
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
@@ -22,23 +22,22 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-package net.runelite.client.plugins.combatnotifier;
+package net.runelite.client.plugins.idlenotifier;
 
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 
 @ConfigGroup(
-	keyName = "combatnotifier",
-	name = "Combat Notifier",
-	description = "Notifies when the player is out of combat"
+	keyName = "idlenotifier",
+	name = "Idle Notifier",
+	description = "Configuration for the idle notifier plugin"
 )
-public interface CombatNotifierConfig
+public interface IdleNotifierConfig
 {
 	@ConfigItem(
 		keyName = "enabled",
 		name = "Enabled",
-		description = "Toggles out of combat notifications"
+		description = "Toggles idle notifications"
 	)
 	default boolean isEnabled()
 	{
@@ -48,10 +47,10 @@ public interface CombatNotifierConfig
 	@ConfigItem(
 		keyName = "timeout",
 		name = "Idle Timeout (ms)",
-		description = "The notification delay after the player is out of combat"
+		description = "The notification delay after the player is idle"
 	)
 	default int getTimeout()
 	{
-		return 10000;
+		return 5000;
 	}
 }
